Changed dump size_t to 4 bytes for compatibility

+++ src/ldump.c
@@ -5,6 +5,7 @@
 */
 
 #include <stddef.h>
+#include <stdint.h>
 
 #define ldump_c
 #define LUA_CORE
@@ -62,12 +63,12 @@ static void DumpString(const TString* s, DumpState* D)
 {
  if (s==NULL || getstr(s)==NULL)
  {
-  size_t size=0;
+  uint32_t size=0;
   DumpVar(size,D);
  }
  else
  {
-  size_t size=s->tsv.len+1;		/* include trailing '\0' */
+  uint32_t size=s->tsv.len+1;		/* include trailing '\0' */
   DumpVar(size,D);
   DumpBlock(getstr(s),size,D);
  }
